refactor existing code
