+++ db/go.mod
@@ -1,0 +1,3 @@
+module github.com/db
+
+go 1.15
